requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.10
filelock==3.13.3
fonttools==4.51.0
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
jmespath==0.10.0
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mmcv==2.1.0
-e git+https://github.com/open-mmlab/mmdetection.git@cfd5d3a985b0249de009b67d04f37263e11cdf3d#egg=mmdet
mmengine==0.10.3
model-index==0.1.11
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.37
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
pandas==2.0.3
pillow==10.3.0
platformdirs==4.2.0
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.2
rich==13.4.2
safetensors==0.4.2
scipy==1.10.1
shapely==2.0.3
six==1.16.0
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.65.2
transformers==4.39.3
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
yapf==0.40.2
zipp==3.18.1
